Split a Pipeline into Its Constituents

Split a compound CPO or CPOTrained into a list of its constituent parts.

This is useful for inspection of pipelines, or for possible rearrangements or changes of pipelines. The resulting list can be changed and rebuilt using pipeCPO.

# S3 method for CPOPrimitive
as.list(x, ...)

# S3 method for CPOTrained
as.list(x, ...)

Arguments

x
[CPO | CPOTrained]
The CPO or CPOTrained chain to split apart.

...
[any]
Ignored.

Value

[list of CPO | list of CPOTrained]. The primitive constituents of x.

See also