Log-Transform a Regression Target Variable.

This is a CPOConstructor to be used to create a CPO. It is called like any R function and returns the created CPO.

Log-transforms the regression Task's target variable.

If predict.type is “response” for inversion, the model's prediction is exponentiated.

If predict.type = “se” prediction is performed, the model's prediction is taken as the parameters of a lognormal random variable; the inverted prediction is then mean = exp(mean + se^2 / 2), se = sqrt((exp(se^2) - 1) * exp(2 * mean + se^2)).

It is therefore recommended to use “se” prediction, possibly with the help of cpoResponseFromSE.

cpoLogTrafoRegr(id)

Arguments

id
[character(1)]
id to use as prefix for the CPO's hyperparameters. this must be used to avoid name clashes when composing two CPOs of the same type, or with learners or other CPOS with hyperparameters with clashing names.

Value

[CPO].

General CPO info

This function creates a CPO object, which can be applied to Tasks, data.frames, link{Learner}s and other CPO objects using the %>>% operator.

The parameters of this object can be changed after creation using the function setHyperPars. The other hyper-parameter manipulating functins, getHyperPars and getParamSet similarly work as one expects.

If the “id” parameter is given, the hyperparameters will have this id as aprefix; this will, however, not change the parameters of the creator function.

Calling a CPOConstructor

CPO constructor functions are called with optional values of parameters, and additional “special” optional values. The special optional values are the id parameter, and the affect.* parameters. The affect.* parameters enable the user to control which subset of a given dataset is affected. If no affect.* parameters are given, all data features are affected by default.

See also